Squashed commit of the following:
commit 9e0d87d71dd3fafdd2fb4d30c3ea4cdb52e1849a
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Fri Dec 7 07:18:58 2018 -0600

    update docs, cleanup

commit 1271d3d63edd1b78092fedbe4ea325cd7d806971
Merge: 033ac9ceb f74fc5979
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Fri Dec 7 07:12:49 2018 -0600

    Merge remote-tracking branch 'upstream/master' into ea-where

commit 033ac9ceb00bf5212bec971ca76092584579b89b
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Fri Dec 7 06:30:18 2018 -0600

    Setitem-based where

commit e9665b89b34acda8757a574250bd25d0b925dcec
Merge: 5e144149e 03134cb92
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Thu Dec 6 21:38:42 2018 -0600

    Merge remote-tracking branch 'upstream/master' into ea-where

commit 5e144149ef4dba21df299cf65ed4e4641ee53cdc
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Thu Dec 6 09:18:54 2018 -0600

    where versionadded

commit d90f3849b360b6fe1680a856c8f918226fc16833
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Thu Dec 6 09:17:43 2018 -0600

    deprecation note for categorical

commit 4715ef6094e006b1f69e5854c6a341a389159924
Merge: edff47ee7 b78aa8d85
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Thu Dec 6 08:15:26 2018 -0600

    Merge remote-tracking branch 'upstream/master' into ea-where

commit edff47ee7f84aa86849c7ed3fc39969319ed7ea4
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Thu Dec 6 08:15:21 2018 -0600

    32-bit compat

commit badb5be34672888fcb1267ad4254b14ebb94d447
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Thu Dec 6 06:21:44 2018 -0600

    compat, revert

commit 911a2daf74d20290965b80159793911f2c2d6249
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Wed Dec 5 15:55:24 2018 -0600

    debug 32-bit issue

commit a69dbb3c31cf6007cae3be629855f381703eafe4
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Wed Dec 5 15:49:17 2018 -0600

    warn for categorical

commit 6f79282127d923deaa30ef0bcb703e5d89d243a7
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Wed Dec 5 12:45:54 2018 -0600

    32-bit compat

commit 56470c31a71e5f416ab3b9af21347781209432cb
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Wed Dec 5 11:39:48 2018 -0600

    Fixups:

    * Ensure data generated OK.
    * Remove erroneous comments about alignment. That was user error.

commit c4604df534abb15a9127adb887066e35dc16a3cc
Author: Tom Augspurger <tom.w.augspurger@gmail.com>
Date:   Mon Dec 3 14:23:25 2018 -0600

    API: Added ExtensionArray.where

    We need some way to do `.where` on EA object for DatetimeArray. Adding it
    to the interface is, I think, the easiest way.

    Initially I started to write a version on ExtensionBlock, but it proved
    to be unwieldy. to write a version that performed well for all types.
    It *may* be possible to do using `_ndarray_values` but we'd need a few more
    things around that (missing values, converting an arbitrary array to the
    "same' ndarary_values, error handling, re-constructing). It seemed easier
    to push this down to the array.

    The implementation on ExtensionArray is readable, but likely slow since
    it'll involve a conversion to object-dtype.

    Closes #24077

+++ pandas/tests/extension/test_sparse.py
@@ -12,7 +12,7 @@ def make_data(fill_value):
     if np.isnan(fill_value):
         data = np.random.uniform(size=100).astype('float64')
     else:
-        data = np.random.randint(1, 100, size=100, dtype='int64')
+        data = np.random.randint(1, 100, size=100)
         if data[0] == data[1]:
             data[0] += 1
 
@@ -266,13 +266,13 @@ class TestMethods(BaseSparseTests, base.BaseMethodsTests):
 
         cond = np.array([True, True, False, False])
         result = ser.where(cond)
-        # new_dtype is the only difference
+
         new_dtype = SparseDtype('float', 0.0)
         expected = pd.Series(cls._from_sequence([a, a, na_value, na_value],
                                                 dtype=new_dtype))
         self.assert_series_equal(result, expected)
 
-        other = cls._from_sequence([a, b, a, b])
+        other = cls._from_sequence([a, b, a, b], dtype=data.dtype)
         cond = np.array([True, False, True, True])
         result = ser.where(cond, other)
         expected = pd.Series(cls._from_sequence([a, b, b, b],
